Refactor & clean up environment variables
