+++ web/.env.cypress
@@ -1,11 +1,3 @@
-REACT_APP_API_BASE_URL=http://non-exist-host/api
 REACT_APP_HIDE_DISCLAIMER=false
-REACT_APP_FIDER_LINK=https://psufiderdev.pathfinder.gov.bc.ca/
-REACT_APP_KEYCLOAK_AUTH_URL=https://sso-dev.pathfinder.gov.bc.ca/auth
-REACT_APP_KEYCLOAK_REALM=8wl6x4cp
-REACT_APP_KEYCLOAK_CLIENT=wps-web
-REACT_APP_MATOMO_URL=matomo-auzhsi-dev.pathfinder.gov.bc.ca
-REACT_APP_MATOMO_SITE_ID=1
-REACT_APP_MATOMO_CONTAINER=kadhkvyA
 BROWSER=none
 PORT=3030
