+++ web/src/api/axios.ts
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { API_BASE_URL } from 'utils/constants'
+import { API_BASE_URL } from 'utils/env'
 
 const instance = axios.create({
   baseURL: API_BASE_URL
